test: split native command compatibility suite
